+++ src/api/generic_device_handler.rs
@@ -33,20 +33,20 @@ impl GenericDeviceHandler {
         self.client.set_device_info(json).await
     }
 
-    /// Gets *device info* as [`GenericDeviceInfoResult`].
+    /// Returns *device info* as [`GenericDeviceInfoResult`].
     /// It is not guaranteed to contain all the properties returned from the Tapo API.
     /// If the deserialization fails, or if a property that you care about it's not present, try [`GenericDeviceHandler::get_device_info_json`].
     pub async fn get_device_info(&self) -> Result<GenericDeviceInfoResult, Error> {
         self.client.get_device_info().await
     }
 
-    /// Gets *device info* as [`serde_json::Value`].
+    /// Returns *device info* as [`serde_json::Value`].
     /// It contains all the properties returned from the Tapo API.
     pub async fn get_device_info_json(&self) -> Result<serde_json::Value, Error> {
         self.client.get_device_info().await
     }
 
-    /// Gets *device usage* as [`DeviceUsageResult`].
+    /// Returns *device usage* as [`DeviceUsageResult`].
     pub async fn get_device_usage(&self) -> Result<DeviceUsageResult, Error> {
         self.client.get_device_usage().await
     }